{{cookiecutter.project_name}}/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest==7.1.3
pytest-flask
pytest-cov==4.0.0
WebTest==3.0.0
factory-boy==3.2.1
pdbpp==0.10.3

# Lint and code style
black==22.10.0
flake8-blind-except==0.2.1
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-isort==4.2.0
flake8==5.0.4
isort==5.10.1
pep8-naming==0.13.2
yapf
mypy
